+++ src/main/java/DNar.java
@@ -1,0 +1,7 @@
+public class DNar {
+    public static void main(String[] args) {
+        UI ui = new UI();
+        ui.greet();
+        ui.exit();
+    }
+}
