+++ src/main/java/UI.java
@@ -1,0 +1,16 @@
+public class UI {
+    public void printLine() {
+        System.out.println("____________________________________________________________");
+    }
+    public void greet() {
+        printLine();
+        System.out.println(" Yo! I'm DNar");
+        System.out.println(" Can you?");
+        printLine();
+    }
+    public void exit() {
+        printLine();
+        System.out.println(" Bye. Don't come back!");
+        printLine();
+    }
+}
